+++ core/libp2p/crypto/hmac_provider/hmac_provider.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "libp2p/crypto/hmac/hmac_provider.hpp"
+#include "libp2p/crypto/hmac_provider/hmac_provider.hpp"
 
 #include <openssl/hmac.h>
 #include "libp2p/crypto/error.hpp"
