+++ core/libp2p/crypto/random_generator.hpp
@@ -6,7 +6,7 @@
 #ifndef KAGOME_CORE_LIBP2P_CRYPTO_RANDOM_RANDOM_GENERATOR_HPP
 #define KAGOME_CORE_LIBP2P_CRYPTO_RANDOM_RANDOM_GENERATOR_HPP
 
-#include "common/buffer.hpp"
+#include "../../common/buffer.hpp"
 
 namespace libp2p::crypto::random {
 
@@ -14,6 +14,7 @@ namespace libp2p::crypto::random {
   class RandomGenerator {
    protected:
     using Buffer = kagome::common::Buffer;
+
    public:
     virtual ~RandomGenerator() = default;
     /**
@@ -22,6 +23,27 @@ namespace libp2p::crypto::random {
      * @return buffer containing random bytes
      */
     virtual Buffer randomBytes(size_t len) = 0;
+  };
+
+  /**
+   * @class CSPRNG provides interface to cryptographic-secure random bytes
+   * generator
+   */
+  class CSPRNG : public RandomGenerator {
+   public:
+    ~CSPRNG() override = default;
+
+    Buffer randomBytes(size_t len) override = 0;
+  };
+
+  /**
+   * @class PseudoRandomGenerator provides interface for pseudo-random generator
+   */
+  class PRNG : public RandomGenerator {
+   public:
+    ~PRNG() override = default;
+
+    Buffer randomBytes(size_t len) override = 0;
   };
 }  // namespace libp2p::crypto::random
