+++ core/libp2p/crypto/random_generator/CMakeLists.txt
@@ -1,0 +1,16 @@
+#
+# Copyright Soramitsu Co., Ltd. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+add_library(random_generator
+    boost_generator.hpp
+    boost_generator.cpp
+    std_generator.hpp
+    std_generator.cpp
+    )
+
+target_link_libraries(random_generator
+    buffer
+    Boost::random
+    )
