+++ core/libp2p/crypto/random_generator/std_generator.hpp
@@ -8,7 +8,7 @@
 
 #include <random>
 
-#include "libp2p/crypto/random/prng.hpp"
+#include "libp2p/crypto/random_generator.hpp"
 
 namespace libp2p::crypto::random {
   /**
