+++ test/core/libp2p/crypto/hmac_test.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "libp2p/crypto/hmac/hmac_provider.hpp"
+#include "libp2p/crypto/hmac_provider/hmac_provider.hpp"
 
 #include <gtest/gtest.h>
 #include <outcome/outcome.hpp>
